Added Multiple Lights

+++ src/Descriptors.cpp
@@ -181,7 +181,7 @@ namespace Aura {
         for (auto& write : writes) {
             write.dstSet = set;
         }
-        vkUpdateDescriptorSets(pool.device.device(), writes.size(), writes.data(), 0, nullptr);
+        vkUpdateDescriptorSets(pool.device.device(), static_cast<uint32_t>(writes.size()), writes.data(), 0, nullptr);
     }
 
 }
